Add browser end2end tests
