+++ package.json
@@ -30,7 +30,10 @@
     "dist:prepare": "run-s dist:clean dist:create",
     "dist:create": "mkdir dist",
     "dist:clean": "rimraf dist",
-    "test": "jest",
+    "test": "run-s test:unit",
+    "test:unit": "cross-env TEST_UNIT=true jest",
+    "test:e2e": "cross-env TEST_E2E=true jest",
+    "test:all": "jest",
     "prepare": "is-ci || husky install"
   },
   "author": {
@@ -44,20 +47,25 @@
   },
   "devDependencies": {
     "@types/jest": "27.5.0",
+    "@types/jest-environment-puppeteer": "5.0.2",
+    "@types/puppeteer": "5.4.6",
     "@types/svg-path-bounding-box": "1.0.0",
-    "@typescript-eslint/eslint-plugin": "5.22.0",
-    "@typescript-eslint/parser": "5.22.0",
+    "@typescript-eslint/eslint-plugin": "5.23.0",
+    "@typescript-eslint/parser": "5.23.0",
+    "anywhere": "1.6.0",
     "coveralls": "3.1.1",
-    "eslint": "8.14.0",
+    "cross-env": "7.0.3",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-prettier": "4.0.0",
-    "husky": "7.0.4",
+    "husky": "8.0.1",
     "is-ci": "3.0.1",
-    "jest": "28.0.3",
+    "jest": "28.1.0",
+    "jest-puppeteer": "6.1.0",
     "npm-run-all": "4.1.5",
     "prettier": "2.6.2",
     "rimraf": "3.0.2",
-    "simple-icons": "6.20.0",
+    "simple-icons": "6.21.0",
     "svg-path-bounding-box": "1.0.4",
     "ts-jest": "28.0.2",
     "ts-node": "10.7.0",
